+++ date/parse.go
@@ -11,11 +11,11 @@ import (
 )
 
 var (
-	// MaxTextLength allows limiting DefaultParser input.
+	// MaxInputLength allows limiting DefaultParser input.
 	// Set 0 to disable this setting.
 	// ErrInputTooLong is wrapped and used if limit is exceeded.
-	// Note: For year > 9999, increment MaxTextLength to > 10.
-	MaxTextLength = 10
+	// Note: For year > 9999, increment MaxInputLength to > 10.
+	MaxInputLength = 10
 
 	// Parser is used by Date.UnmarshalText function.
 	Parser = DefaultParser
@@ -35,27 +35,27 @@ const (
 
 // DefaultParser parse Date from input.
 //
-// See also MaxTextLength.
-func DefaultParser(data []byte, r Rule) (date Date, err error) {
+// See also MaxInputLength.
+func DefaultParser(input []byte, r Rule) (date Date, err error) {
 	const funcName = "DefaultParser"
-	l := len(data)
+	l := len(input)
 	if l == 0 {
 		return Date{}, newParseError(funcName, "", nil)
 	}
-	if MaxTextLength != 0 && l > MaxTextLength {
+	if MaxInputLength != 0 && l > MaxInputLength {
 		// do not use input for "input too long" error
-		return Date{}, newParseError(funcName, "", fmt.Errorf("%w: %d > %d", ErrInputTooLong, l, MaxTextLength))
+		return Date{}, newParseError(funcName, "", fmt.Errorf("%w: %d > %d", ErrInputTooLong, l, MaxInputLength))
 	}
-	parts := pattern.FindSubmatch(data)
+	parts := pattern.FindSubmatch(input)
 	if len(parts) == 0 {
-		return Date{}, newParseError(funcName, string(data), nil)
+		return Date{}, newParseError(funcName, string(input), nil)
 	}
-	if sep2 := data[l-3] == '-'; sep2 || data[l-5] == '-' { // extended format
-		if !sep2 || data[l-6] != '-' { // disallow YYYY-MMDD and YYYYMM-YY formats
-			return Date{}, newParseError(funcName, string(data), nil)
+	if sep2 := input[l-3] == '-'; sep2 || input[l-5] == '-' { // extended format
+		if !sep2 || input[l-6] != '-' { // disallow YYYY-MMDD and YYYYMM-YY formats
+			return Date{}, newParseError(funcName, string(input), nil)
 		}
 	} else if r&RuleDisableBasic != 0 {
-		return Date{}, newParseError(funcName, string(data), ErrBasicFormatDisabled)
+		return Date{}, newParseError(funcName, string(input), ErrBasicFormatDisabled)
 	}
 	year, _ := strconv.Atoi(string(parts[1]))
 	month, _ := strconv.Atoi(string(parts[2]))