+++ size/size_test.go
@@ -220,8 +220,8 @@ func Test_Size_UnmarshalText(t *testing.T) {
 	value := Size(20 * 1024)
 
 	s := Size(10)
-	Parser = func(data []byte, r Rule) (Size, error) {
-		assert.Equal(t, text, data)
+	Parser = func(input []byte, r Rule) (Size, error) {
+		assert.Equal(t, text, input)
 		assert.Equal(t, DefaultRule&ruleUnmarshalTextMask, r)
 		return value, nil
 	}
@@ -230,8 +230,8 @@ func Test_Size_UnmarshalText(t *testing.T) {
 
 	s = Size(10)
 	parseError := errors.New("parse error")
-	Parser = func(data []byte, r Rule) (Size, error) {
-		assert.Equal(t, text, data)
+	Parser = func(input []byte, r Rule) (Size, error) {
+		assert.Equal(t, text, input)
 		assert.Equal(t, DefaultRule&ruleUnmarshalTextMask, r)
 		return 0, parseError
 	}
@@ -278,8 +278,8 @@ func Test_Size_UnmarshalJSON(t *testing.T) {
 	value := Size(20 * 1024)
 
 	s := Size(10)
-	Parser = func(data []byte, r Rule) (Size, error) {
-		assert.Equal(t, json, data)
+	Parser = func(input []byte, r Rule) (Size, error) {
+		assert.Equal(t, json, input)
 		assert.Equal(t, DefaultRule, r)
 		return value, nil
 	}
@@ -288,8 +288,8 @@ func Test_Size_UnmarshalJSON(t *testing.T) {
 
 	s = Size(10)
 	parseError := errors.New("parse error")
-	Parser = func(data []byte, r Rule) (Size, error) {
-		assert.Equal(t, json, data)
+	Parser = func(input []byte, r Rule) (Size, error) {
+		assert.Equal(t, json, input)
 		assert.Equal(t, DefaultRule, r)
 		return 0, parseError
 	}